Character page work
Updated the server-side character model to support the final selection
of fields (e.g. skills, attributes, level, devotion, slots, etc.).

Updated the per-character page to render out all of the character's
fields (generic style, so far).

Added in a bunch of util methods to initialize and display slots a la
the old example, and copied over the style sheets. It renders! So far
only the "clickable" and "active" tags are wired up, but there are hook
points which should make it straightforward to begin layering on the
rest of the manipulation elements.