Refactored character components
In prep for eventually duplicating all the Character code to support
Monsters (NPCs), I've moved most of the work out of the character,
characters, and home controllers and into the character_support.js
class.

The same thing still needs to happen on the skilltree, which is
currently tied tightly in a few select spots to the Characters database.