Minor fixes
Fixed a null error when you were logged out; turns out Meteor.userId is
a function, not a value.

Also fixed the link from the home page tile so that it doesn't lead to
an uppercase'd URL of the character's name.

Cleaned up the header stylings to be more consistent.